webdriverio: refactor shadowRoot command to use element.$ rename shadowRoot -> shadow$ implement shadow$$
